% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dolores.R
\name{Dolores}
\alias{Dolores}
\title{Dolores}
\usage{
Dolores(phy, states, k, q.div)
}
\arguments{
\item{phy}{a phylogenetic tree, in ape "phylo" format.}

\item{states}{a numeric vector from Forest.knolls() depicting the character states of interest.}

\item{k}{the numeric value of character states. Three is the max number of traits for this model.}

\item{q.div}{the ratio of diversification rate to character change rate. Used to inform the starting point of the MuSSE model.}
}
\description{
Multi-State Speciation and Extinction (MuSSE) model using bayesian based inference on parameters. Speciation and extinction parameters are informed through a population ecology and population genetics framework.
}
\examples{
pars <- c(.1,  .15,  .2,  # lambda 1, 2, 3
        .03, .045, .06,   # mu 1, 2, 3
        .05,    0,        # q12, q13
        .05,  .05,        # q21, q23
         0,   .05)        # q31, q32
set.seed(2)
phy <- tree.musse(pars, 30, x0=1)
K<-1000
p<-1/(2*K)
ntaxa<-204
gens<-100
r<-0.5
states<-Forest.knolls(p,K,r,ntaxa,gens)
k<-3
q.div<-5
Dolores(phy,states,k,q.div)

}
\references{
Beaulieu, J. M., & O’meara, B. C. (2016). Detecting hidden diversification shifts in models of trait-dependent speciation and extinction. Systematic biology, 65(4), 583-601.

FitzJohn, R. G. (2012). Diversitree: comparative phylogenetic analyses of diversification in R. Methods in Ecology and Evolution, 3(6), 1084-1092.

Maddison, W. P., Midford, P. E., & Otto, S. P. (2007). Estimating a binary character's effect on speciation and extinction. Systematic biology, 56(5), 701-710.

Magallon S. and Sanderson M.J. 2001. Absolute diversification rates in angiospem clades. Evol. 55:1762-1780.

O'Meara, B. C., Smith, S. D., Armbruster, W. S., Harder, L. D., Hardy, C. R., Hileman, L. C., ... & Stevens, P. F. (2016). Non-equilibrium dynamics and floral trait interactions shape extant angiosperm diversity. Proc. R. Soc. B, 283(1830), 20152304.
}
\author{
Jonathan R. Dickey
}
