% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twin.peaks.R
\name{Twin.peaks}
\alias{Twin.peaks}
\title{Twin.peaks}
\usage{
Twin.peaks(p,K,r,gens)
}
\arguments{
\item{p}{an object describing the initial frequency of A}

\item{K}{describes the carrying capacity}

\item{r}{a number giving the population growth rate}

\item{gens}{is the number of generations}
}
\description{
Twin.peaks, a function describing selection, survival and reproduction, on a population given a single climate change event that affects temperature in an island ecosystem, 2 degree threshold. Exploring with the geontypes AA, Aa, and aa. No allele has dominance and the population is in hardy weinberg equilibrium to denote genotype frequency.
}
\examples{
p<-1/(2*K)
K<-1000
Twin.peaks(p,K,2,100)
}
